Fix: target 자료형 불일치 수정(#24) -int형에서 double형으로 수정
Resolves: #24

+++ app/src/main/java/com/money/app/module/user/dto/UserDto.java
@@ -26,7 +26,7 @@ public class UserDto { //내부 로직에 사용
     private String job;
     private LocalDateTime createDatetime;
     private LocalDateTime updateDatetime;
-    private int target;
+    private double target;
     private String prefer;
 
     public static UserDto fromEntity(User user) {
